Make HeuristicAtStop a record and move to internalapi
